Move DBus errors to a subpackage
DBus errors shouldn't be in one file. However, we should be careful,
because the client will not recognize errors from files that were not
imported.
