+++ pyanaconda/dbus/errors/Makefile.am
@@ -1,6 +1,6 @@
-# dbus/Makefile.am for anaconda
+# Makefile.am for anaconda
 #
-# Copyright (C) 2017  Red Hat, Inc.
+# Copyright (C) 2018  Red Hat, Inc.
 #
 # This program is free software; you can redistribute it and/or modify
 # it under the terms of the GNU Lesser General Public License as published
@@ -15,10 +15,8 @@
 # You should have received a copy of the GNU Lesser General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-SUBDIRS = errors
-
 pkgpyexecdir = $(pyexecdir)/py$(PACKAGE_NAME)
-dbusdir = $(pkgpyexecdir)/dbus
-dbus_PYTHON = $(srcdir)/*.py
+dbus_errorsdir = $(pkgpyexecdir)/dbus/errors
+dbus_errors_PYTHON = $(srcdir)/*.py
 
 MAINTAINERCLEANFILES = Makefile.in
